+++ data/src/main/java/com/example/data/datasources/api/PhotosFeedApi.kt
@@ -6,5 +6,5 @@ import retrofit2.Response
 interface PhotosFeedApi {
     suspend fun getPhotosFeed(pageIndex: Int, pageSize: Int): Response<PhotosResponse>
 
-    suspend fun getUserPhotos(userId: Int, pageIndex: Int, pageSize: Int): Response<PhotosResponse>
+    suspend fun getUserPhotos(userId: Long, pageIndex: Int, pageSize: Int): Response<PhotosResponse>
 }
